Introduce Config and NewDownloader
Refactor downloader to use a Config struct with validation and defaults.
Add NewDownloader to build an HTTP client and store config/client on the
Downloader. Move flag parsing into parseFlags and update main to use the
new constructor. Replace GetRangeHeaders with RangeHeaders and rename
the ranges type. Remove utils.go and change Makefile default DIST to
./data/file.png.